+++ packages/hono-takibi/src/core/schema/references/get-ref-name.ts
@@ -30,10 +30,10 @@
  * Process: Splits by '/' -> ['#', 'components', 'schemas', 'Address']
  * Output: Returns 'Address'
  */
-export function getRefName(ref: string): string | undefined {
+export function getRefName($ref: `#/components/schemas/${string}`): string | undefined {
   // split('/'): Split a string into an array using slashes
   // 1. ["#", "components", "schemas", "Address"]
   // pop() to get the last element
   // 2. "Address"
-  return ref.split('/').pop()
+  return $ref.split('/').pop()
 }
