+++ packages/hono-takibi/src/core/schema/references/get-ref-schema-name.ts
@@ -1,19 +1,15 @@
-import type { Schema } from '../../../openapi/index.js'
 import { getRefName } from './index.js'
 
 /**
  * Retrieves the referenced schema name from a schema object.
- * @param { Schema } schema - The schema object
+ * @param { `#/components/schemas/${string}` } ref - The reference string in the format `#/components/schemas/{schemaName}`
  * @returns { string } The referenced schema name
  * @throws { Error } Will throw an error if the $ref property or the reference name is not found
  */
-export function getRefSchemaName(schema: Schema): string {
-  if (!schema.$ref) {
+export function getRefSchemaName($ref: `#/components/schemas/${string}`): string {
+  if (!$ref) {
     throw new Error('refName is not found')
   }
-  const refName = getRefName(schema.$ref)
-  if (!refName) {
-    throw new Error('refName is not found')
-  }
+  const refName = getRefName($ref)
   return `${refName}Schema`
 }
