+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/components/anyof/index.test.ts
@@ -2,10 +2,22 @@ import { describe, it, expect } from 'vitest'
 import { anyOf } from '.'
 
 // Test run
-// pnpm vitest run ./src/generator/zod-openapi-hono/openapi/component/anyOf/index.test.ts
+// pnpm vitest run ./src/generator/zod-openapi-hono/openapi/component/anyof/index.test.ts
 
-describe('anyOf barrel file exports', () => {
-  it.concurrent('should export anyOf function', () => {
-    expect(typeof anyOf).toBe('function')
+describe('generateAnyOfCode', () => {
+  it.concurrent('generateAnyOfCode with anyOf', () => {
+    const result = anyOf({
+      anyOf: [
+        {
+          $ref: '#/components/schemas/MultiPolygon',
+        },
+        {
+          $ref: '#/components/schemas/Polygon',
+        },
+      ],
+      description: 'Center coordinates',
+    })
+    const expected = 'z.union([MultiPolygonSchema,PolygonSchema])'
+    expect(result).toBe(expected)
   })
 })
