+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/components/not/index.test.ts
@@ -4,8 +4,10 @@ import { not } from '.'
 // Test run
 // pnpm vitest run ./src/generator/zod-openapi-hono/openapi/component/not/index.test.ts
 
-describe('not barrel file exports', () => {
-  it.concurrent('should export not function', () => {
-    expect(typeof not).toBe('function')
+describe('not', () => {
+  it.concurrent(`not({ not: { type: 'number' } }) -> 'z.unknown()'`, () => {
+    const result = not({ not: { type: 'number' } })
+    const expected = 'z.unknown()'
+    expect(result).toBe(expected)
   })
 })
