--- packages/hono-takibi/src/generator/zod-openapi-hono/openapi/components/not/not.test.ts
@@ -1,13 +1,0 @@
-import { describe, it, expect } from 'vitest'
-import { not } from '.'
-
-// Test run
-// pnpm vitest run ./src/generator/zod-openapi-hono/openapi/component/not/not.test.ts
-
-describe('not', () => {
-  it.concurrent(`not({ not: { type: 'number' } }) -> 'z.unknown()'`, () => {
-    const result = not({ not: { type: 'number' } })
-    const expected = 'z.unknown()'
-    expect(result).toBe(expected)
-  })
-})
