+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/components/oneof/index.test.ts
@@ -4,8 +4,20 @@ import { oneOf } from '.'
 // Test run
 // pnpm vitest run ./src/generator/zod-openapi-hono/openapi/component/oneof/index.test.ts
 
-describe('oneOf barrel file exports', () => {
-  it.concurrent('should export oneOf function', () => {
-    expect(typeof oneOf).toBe('function')
+describe('oneOf', () => {
+  it.concurrent('oneOf with oneOf', () => {
+    const result = oneOf({
+      oneOf: [
+        {
+          type: 'number',
+        },
+        {
+          type: 'string',
+        },
+      ],
+    })
+
+    const expected = 'z.union([z.number(),z.string()])'
+    expect(result).toBe(expected)
   })
 })
