+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/create-route.test.ts
@@ -8,12 +8,12 @@ describe('createRoute', () => {
   it.concurrent('createRoute Test', () => {
     const result = createRoute({
       routeName: 'deletePostsId',
-      tagsCode: '["Post"]',
-      methodCode: 'delete',
-      pathCode: '/posts/{id}',
-      descriptionCode: 'delete post',
+      tags: '["Post"]',
+      method: 'delete',
+      path: '/posts/{id}',
+      description: 'delete post',
       requestParams: 'request:{params:z.object({id:z.string().uuid()})},',
-      responsesCode: `204:{description:'No Content',},400:{description:'Bad Request',content:{'application/json':{schema:z.object({message:z.string()}),},},},500:{description:'Internal Server Error',content:{'application/json':{schema: z.object({message: z.string()}),},},},`,
+      responses: `204:{description:'No Content',},400:{description:'Bad Request',content:{'application/json':{schema:z.object({message:z.string()}),},},},500:{description:'Internal Server Error',content:{'application/json':{schema: z.object({message: z.string()}),},},},`,
     })
     const expected = `export const deletePostsId=createRoute({["Post"]delete/posts/{id}delete postrequest:{params:z.object({id:z.string().uuid()})},204:{description:'No Content',},400:{description:'Bad Request',content:{'application/json':{schema:z.object({message:z.string()}),},},},500:{description:'Internal Server Error',content:{'application/json':{schema: z.object({message: z.string()}),},},},})`
     expect(result).toBe(expected)
